Remove Device: Used to distribute all the bricks from device to other devices
Important:
We only support this feature for Replicate and Distribute Replicate
Volume.
Refer:
https://gluster.readthedocs.io/en/latest/Administrator%20Guide/Managing%20Volumes/#replace-brick

Removes all the brick from the device and start these bricks on other
devices based on allocator. Heal is triggered automatially for replicate
volumes
on replace brick.

Allocate and create new brick to replace.
It stops the brick to be replaced, If it is not already down(kill the
brick process).
Then gluster replace brick which will replace the brick with new one and
also starts the heals.

If other nodes does not have sufficient storage then this command should
fail.

Pair-Programmed-With: Mohamed Ashiq Liyazudeen <mliyazud@redhat.com>
Signed-off-by: Mohamed Ashiq Liyazudeen mliyazud@redhat.com
Signed-off-by: Raghavendra Talur <rtalur@redhat.com>